app/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
psycopg2-binary==2.9.9
pytest==8.3.2
httpx==0.27.0
python-dotenv==1.0.1
pytest-asyncio==0.24.0